- Progress with octomaps. - Fixed a pair of serialization bugs in rarely used classes (detected with new unit tests). - TColor & TColorf are now serializable.

+++ libs/base/src/poses/CPointPDFSOG.cpp
@@ -187,9 +187,9 @@ void  CPointPDFSOG::readFromStream(CStream &in,int version)
 	case 0:
 	case 1:
 		{
-			uint32_t								N;
+			uint32_t							N;
 			CListGaussianModes::iterator		it;
-			float									x;
+			double								x;
 
 			in >> N;
